+++ app/page.tsx
@@ -1,6 +1,11 @@
+'use client'
+
 import Matrice from "@/components/2048";
+import { GameContextProvider } from "@/components/2048/context";
 
 export default function Home() {
 
-  return <Matrice />;
+  return <GameContextProvider>
+    <Matrice />
+  </GameContextProvider>
 }
